bpf: initial datapath acceleration for nodeport in xdp
Rework the XDP prefilter early drop to now also include NodePort service
handling. Idea is that after prefilter phase, we can accelerate the case
where the NodePort(/ExternalIP) backend is remote and we are forced to
push the packet back out again via SNAT or DSR.

Instead of first crafting an skb, pushing it through GRO engine, packet
taps to eventually end up at tc ingress of the phy device, we can do all
this right in the driver's XDP layer instead at a much higher per packet
rate. This is the case where XDP can really help to help scaling up the
backends. XDP won't be able to help much for the case where the backend
is local to the node, and when we're hitting cases where we need to punt
to remote backends, then can can do so at max the rate that tc ingress
can handle, so moving this down into XDP will improve that limit. At
the same time it will also create less burden given unneeded layers are
bypassed.

All major 10/40G NICs these days support XDP in the upstream kernel. This
is also in particular useful e.g. in cloud-based environments with SRIO-V
networking or on bare metal machines in general.

This work piggy-back on the entire datapath conversion that has been done
for Cilium's BPF code in order to be generic for skb and xdp.

We add daemon config flag, so we can opt-into --node-port-acceleration={none,
generic,driver} setting in order to enable NodePort via XDP. The default
is 'none' which will then use the regular tc ingress based mechanism. So
far it's opt-in, at some point we can think about integrating it into the
kube-proxy-free probe mode.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>