test: Extend testNodePort
Extend the test function to accept a param which indicates whether
tests from a third (aka node w/o cilium) node should be performed.

As a reminder, the third node is used for NodePort testing to avoid
bpf_sock.c optimizations which happen when a request is sent to
a NodePort svc from a node which runs cilium.

Signed-off-by: Martynas Pumputis <m@lambda.lt>